Remove Progress Handler, Move Logic To ProgressModel
